+++ app/src/main/java/cn/wthee/pcrtool/ui/character/CharacterPagerFragment.kt
@@ -1,7 +1,6 @@
 package cn.wthee.pcrtool.ui.character
 
 import android.content.Context
-import android.content.res.ColorStateList
 import android.graphics.Color
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -21,7 +20,6 @@ import cn.wthee.pcrtool.databinding.FragmentCharacterPagerBinding
 import cn.wthee.pcrtool.ui.character.CharacterPagerFragment.Companion.uid
 import cn.wthee.pcrtool.ui.character.attr.CharacterAttrFragment
 import cn.wthee.pcrtool.ui.character.attr.CharacterDropDialogFragment
-import cn.wthee.pcrtool.ui.home.CharacterListFragment
 import cn.wthee.pcrtool.ui.skill.SkillFragment
 import cn.wthee.pcrtool.ui.skill.SkillLoopDialogFragment
 import cn.wthee.pcrtool.utils.Constants
@@ -51,7 +49,6 @@ class CharacterPagerFragment : Fragment() {
         lateinit var viewPager: ViewPager2
         var uid = -1
         var currentPage = 0
-        var isLoved = false
     }
 
 
@@ -74,7 +71,6 @@ class CharacterPagerFragment : Fragment() {
             name = it.getString(UNIT_NAME) ?: ""
             nameEx = it.getString(Constants.UNIT_NAME_EX) ?: ""
         }
-        isLoved = CharacterListFragment.characterFilterParams.starIds.contains(uid)
         sharedElementEnterTransition = MaterialContainerTransform().apply {
             scrimColor = Color.TRANSPARENT
             duration = 500L
@@ -94,8 +90,6 @@ class CharacterPagerFragment : Fragment() {
         init()
         //角色图片列表
         setListener()
-        //初始收藏
-        setLove(isLoved)
         return binding.root
     }
 
@@ -120,18 +114,6 @@ class CharacterPagerFragment : Fragment() {
                 null,
                 extras
             )
-        }
-
-        //收藏点击监听
-        binding.fabLove.setOnClickListener {
-            isLoved = !isLoved
-            CharacterListFragment.characterFilterParams.addOrRemove(uid)
-            setLove(isLoved)
-        }
-        binding.btnLove.setOnClickListener {
-            isLoved = !isLoved
-            CharacterListFragment.characterFilterParams.addOrRemove(uid)
-            setLove(isLoved)
         }
     }
 
@@ -268,19 +250,5 @@ class CharacterPagerFragment : Fragment() {
             R.dimen.viewpager_current_item_horizontal_margin
         )
         this.addItemDecoration(itemDecoration)
-    }
-
-    /**
-     * 更新收藏按钮颜色
-     */
-    private fun setLove(isLoved: Boolean) {
-        val icFabColor =
-            ResourcesUtil.getColor(if (isLoved) R.color.colorPrimary else R.color.alphaPrimary)
-
-        val drawable =
-            ResourcesUtil.getDrawable(if (isLoved) R.drawable.ic_loved else R.drawable.ic_loved_line)
-
-        binding.fabLove.imageTintList = ColorStateList.valueOf(icFabColor)
-        binding.btnLove.setImageDrawable(drawable)
     }
 }
