+++ app/src/main/res/layout/item_skill.xml
@@ -7,7 +7,7 @@
 
     <com.google.android.material.textview.MaterialTextView
         android:id="@+id/name"
-        android:layout_width="0dp"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_margin="@dimen/s_dp"
         android:textAppearance="@style/TextAppearance.MaterialComponents.Subtitle2"
@@ -15,30 +15,16 @@
         android:textIsSelectable="true"
         android:textStyle="bold"
         app:layout_constrainedWidth="true"
-        app:layout_constraintEnd_toStartOf="@+id/type"
+        app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_weight="1"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        tools:text="@string/test" />
+        tools:text="@string/app_name" />
 
-    <com.google.android.material.textview.MaterialTextView
-        android:id="@+id/type"
-        style="@style/TitleStyle"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_margin="@dimen/s_dp"
-        app:layout_constraintBottom_toBottomOf="@id/name"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="@id/name"
-        tools:text="必杀技" />
-
-
-    <LinearLayout
+    <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/layout_skill_desc"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:nestedScrollingEnabled="false"
-        android:orientation="horizontal"
         android:padding="@dimen/s_dp"
         app:layout_constraintTop_toBottomOf="@id/name">
 
@@ -47,40 +33,49 @@
             android:layout_width="@dimen/normal_icon"
             android:layout_height="@dimen/normal_icon"
             android:adjustViewBounds="true"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
             tools:src="@drawable/unknown_gray" />
 
-        <LinearLayout
+        <com.google.android.material.textview.MaterialTextView
+            android:id="@+id/type"
+            style="@style/TitleStyle"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            tools:text="必杀技" />
+
+        <com.google.android.material.textview.MaterialTextView
+            android:id="@+id/level"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_marginStart="@dimen/m_dp"
-            android:orientation="vertical">
+            android:gravity="start"
+            android:layout_marginStart="@dimen/s_dp"
+            android:textAppearance="@style/TextAppearance.MaterialComponents.Subtitle2"
+            android:textColor="@color/colorPrimary"
+            android:textStyle="bold"
+            app:layout_constrainedWidth="true"
+            app:layout_constraintStart_toEndOf="@id/item_pic"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintEnd_toStartOf="@id/type"
+            tools:text="999" />
 
-            <com.google.android.material.textview.MaterialTextView
-                android:id="@+id/level"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:gravity="start"
-                android:paddingStart="@dimen/xs_dp"
-                android:paddingEnd="@dimen/xs_dp"
-                android:textAppearance="@style/TextAppearance.MaterialComponents.Subtitle2"
-                android:textColor="@color/colorPrimary"
-                android:textStyle="bold"
-                app:layout_constrainedWidth="true"
-                app:layout_constraintWidth_percent="0.25"
-                tools:text="999" />
+        <com.google.android.material.textview.MaterialTextView
+            android:id="@+id/desc"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:nestedScrollingEnabled="false"
+            android:textAppearance="@style/TextAppearance.MaterialComponents.Subtitle2"
+            android:textIsSelectable="true"
+            app:layout_constrainedWidth="true"
+            android:layout_marginStart="@dimen/s_dp"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toEndOf="@id/item_pic"
+            app:layout_constraintTop_toBottomOf="@id/level"
+            tools:text="@string/test" />
 
-            <com.google.android.material.textview.MaterialTextView
-                android:id="@+id/desc"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:nestedScrollingEnabled="false"
-                android:textAppearance="@style/TextAppearance.MaterialComponents.Subtitle2"
-                android:textIsSelectable="true"
-                tools:text="@string/test" />
-
-        </LinearLayout>
-
-    </LinearLayout>
+    </androidx.constraintlayout.widget.ConstraintLayout>
 
 
     <cn.wthee.pcrtool.utils.NoTouchRecyclerView
