+++ app/src/main/res/layout/layout_equip_detail.xml
@@ -18,43 +18,37 @@
         app:layout_constraintTop_toTopOf="parent" />
 
 
-    <LinearLayout
+    <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/layout_equip"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="@dimen/s_dp"
         android:orientation="horizontal"
+        android:padding="@dimen/s_dp"
         app:layout_constraintTop_toBottomOf="@id/equip_name">
 
 
-        <LinearLayout
+        <androidx.appcompat.widget.AppCompatImageView
+            android:id="@+id/item_pic"
+            android:layout_width="@dimen/normal_icon"
+            android:layout_height="@dimen/normal_icon"
+            android:scaleType="fitCenter"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            tools:src="@drawable/unknown_gray" />
+
+        <com.google.android.material.textview.MaterialTextView
+            android:id="@+id/ue_lv"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginStart="@dimen/m_dp"
-            android:orientation="vertical">
-
-            <androidx.appcompat.widget.AppCompatImageView
-                android:id="@+id/item_pic"
-                android:layout_width="@dimen/normal_icon"
-                android:layout_height="@dimen/normal_icon"
-                android:scaleType="fitCenter"
-                tools:src="@drawable/unknown_gray" />
-
-            <com.google.android.material.textview.MaterialTextView
-                android:id="@+id/ue_lv"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center_horizontal"
-                android:paddingStart="@dimen/xs_dp"
-                android:paddingEnd="@dimen/xs_dp"
-                android:textAppearance="@style/TextAppearance.MaterialComponents.Subtitle2"
-                android:textColor="@color/colorPrimary"
-                android:textStyle="bold"
-                app:layout_constrainedWidth="true"
-                app:layout_constraintWidth_percent="0.25"
-                tools:text="999" />
-
-        </LinearLayout>
+            android:layout_gravity="center_horizontal"
+            android:textAppearance="@style/TextAppearance.MaterialComponents.Subtitle2"
+            android:textColor="@color/colorPrimary"
+            android:textStyle="bold"
+            app:layout_constrainedWidth="true"
+            app:layout_constraintEnd_toEndOf="@id/item_pic"
+            app:layout_constraintStart_toStartOf="@id/item_pic"
+            app:layout_constraintTop_toBottomOf="@id/item_pic"
+            tools:text="999" />
 
         <com.google.android.material.textview.MaterialTextView
             android:id="@+id/desc"
@@ -63,8 +57,12 @@
             android:layout_marginStart="@dimen/s_dp"
             android:textAppearance="@style/TextAppearance.MaterialComponents.Body2"
             android:textIsSelectable="true"
+            app:layout_constrainedWidth="true"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toEndOf="@id/item_pic"
+            app:layout_constraintTop_toTopOf="parent"
             tools:text="@string/test" />
-    </LinearLayout>
+    </androidx.constraintlayout.widget.ConstraintLayout>
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/equip_attrs"
