+++ app/src/main/res/values/strings.xml
@@ -110,7 +110,8 @@
     <string name="tool_gacha">卡池</string>
     <string name="tool_event">活动</string>
     <string name="drop_info">- 掉落信息 -</string>
-    <string name="test">testtesttestesttesttestesttesttestesttesttestesttesttestesttesttes</string>
+    <string name="test">testtesttestesttesttestesttestttesttesttestesttesttestesttesttestesttesttestesttesttestesttesttesestesttesttestesttesttestesttesttes</string>
+    <string name="test_short">test</string>
     <string name="tool_news">公告</string>
     <string name="loading">加载中…</string>
     <string name="loading_news">正在获取最新数据…</string>
