+++ app/src/main/res/xml/fragment_character_pager_scene.xml
@@ -23,61 +23,29 @@
             app:layout_constraintEnd_toEndOf="@id/name"
             app:layout_constraintStart_toStartOf="@id/name" />
 
-        <Constraint
-            android:id="@id/fab_love"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            app:layout_constraintBottom_toBottomOf="@id/character_pic"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/character_pic" />
-
-        <Constraint
-            android:id="@id/btn_love"
-            android:layout_width="@dimen/l_icon"
-            android:layout_height="@dimen/l_icon"
-            app:layout_constraintBottom_toBottomOf="@id/character_pic"
-            app:layout_constraintEnd_toEndOf="@id/fab_love"
-            app:layout_constraintStart_toStartOf="@id/fab_love"
-            app:layout_constraintTop_toBottomOf="@id/character_pic" />
     </ConstraintSet>
 
     <ConstraintSet android:id="@+id/collapsed">
         <Constraint
             android:id="@id/character_pic"
             android:layout_width="match_parent"
-            android:layout_height="@dimen/toolbar_height"
+            android:layout_height="@dimen/toolbar_height_1"
             app:layout_constraintTop_toTopOf="parent" />
 
         <Constraint
             android:id="@id/name"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            app:layout_constraintBottom_toTopOf="@id/name_ex" />
+            app:layout_constraintTop_toTopOf="@id/character_pic" />
 
         <Constraint
             android:id="@id/name_ex"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            app:layout_constraintBottom_toBottomOf="@id/character_pic"
             app:layout_constraintEnd_toEndOf="@id/name"
-            app:layout_constraintStart_toStartOf="@id/name" />
+            app:layout_constraintStart_toStartOf="@id/name"
+            app:layout_constraintTop_toBottomOf="@id/name" />
 
-        <Constraint
-            android:id="@id/fab_love"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            app:layout_constraintBottom_toBottomOf="@id/character_pic"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="@id/character_pic" />
-
-        <Constraint
-            android:id="@id/btn_love"
-            android:layout_width="@dimen/l_icon"
-            android:layout_height="@dimen/l_icon"
-            app:layout_constraintBottom_toBottomOf="@id/character_pic"
-            app:layout_constraintEnd_toEndOf="@id/fab_love"
-            app:layout_constraintStart_toStartOf="@id/fab_love"
-            app:layout_constraintTop_toTopOf="@id/character_pic" />
     </ConstraintSet>
 
     <Transition
@@ -118,24 +86,6 @@
                 android:alpha="1"
                 app:framePosition="100"
                 app:motionTarget="@id/name_ex" />
-
-            <KeyAttribute
-                android:alpha="1.0"
-                app:framePosition="0"
-                app:motionTarget="@id/fab_love" />
-            <KeyAttribute
-                android:alpha="0.0"
-                app:framePosition="100"
-                app:motionTarget="@id/fab_love" />
-
-            <KeyAttribute
-                android:alpha="0"
-                app:framePosition="0"
-                app:motionTarget="@id/btn_love" />
-            <KeyAttribute
-                android:alpha="1"
-                app:framePosition="100"
-                app:motionTarget="@id/btn_love" />
         </KeyFrameSet>
         <OnSwipe
             app:dragDirection="dragUp"
